Marketing Connections: Make the UI aware of the refresh-failed status
Instagram, Facebook and possibly other, tokens need to be refreshed on a
regular basis. If we aren't able to, we should indicate that the
connection will break after the token expiry date.

This is the front end changes to D52030-code, which sets adds the
`refresh-failed` status to the API endpoint. As such this PR should be
merged before the backend changes.

If the connection is returned as having the `refresh-failed` status then
we treat it like a broken connection, displaying the 'Reconnect' button,
and a message about when the connection is due to expire.